requirements.txt
APScheduler==3.6.3
astroid==2.7.3
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
idna==3.2
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
nose==1.3.7
platformdirs==2.3.0
pylint==2.10.2
python-dotenv==0.19.0
python-telegram-bot==13.7
pytz==2021.1
PyYAML==5.4.1
Random-Word==1.0.7
requests==2.26.0
six==1.16.0
toml==0.10.2
tornado==6.1
typed-ast==1.4.3
typing-extensions==3.10.0.2
tzlocal==3.0
urllib3==1.26.6
wrapt==1.12.1
